habit_tracker_app/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary
drf-spectacular